Changed number data types from long to double or int

+++ src/Allegiance.Blazor.Highcharts/Options/Background.cs
@@ -17,7 +17,7 @@ namespace Allegiance.Blazor.Highcharts.Options
         public string InnerRadius { get; set; }
 
         [JsonProperty("borderWidth", NullValueHandling = NullValueHandling.Ignore)]
-        public long? BorderWidth { get; set; }
+        public int? BorderWidth { get; set; }
 
         //////////////////////////////////////////////////////
