+++ src/Allegiance.Blazor.Highcharts/Options/Pane.cs
@@ -6,9 +6,9 @@ namespace Allegiance.Blazor.Highcharts.Options
     public partial class Pane
     {
         [JsonProperty("startAngle", NullValueHandling = NullValueHandling.Ignore)]
-        public long? StartAngle { get; set; }
+        public int? StartAngle { get; set; }
         [JsonProperty("endAngle", NullValueHandling = NullValueHandling.Ignore)]
-        public long? EndAngle { get; set; }
+        public int? EndAngle { get; set; }
 
         [JsonProperty("background", NullValueHandling = NullValueHandling.Ignore)]
         public List<Background> Background { get; set; }
