+++ src/Allegiance.Blazor.Highcharts/Options/Stop.cs
@@ -6,10 +6,10 @@ namespace Allegiance.Blazor.Highcharts.Options
 {
     public partial class Stop
     {
-        public long? Value { get; set; }
+        public double? Value { get; set; }
         public string Color { get; set; }
 
-        public Stop(long? value, string color)
+        public Stop(double? value, string color)
         {
             Value = value;
             Color = color;
